Move R4 project ID to separate constants file
